requirements.txt
streamlit>=1.32.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
ultralytics>=8.0.0
faiss-cpu>=1.7.4
Pillow>=9.0.0
numpy>=1.24.0
pandas>=2.0.0
